update MIME info when building environment
Properly use the pre-defined MIME type associations that come with the
installed packages, so that we don't have to chose the correct default
application the first time we open a corresponding file.

Firefox now also recognises it's the default browser on first start.

This has the added bonus that we don't have to track `mimeinfo.list`,
which is a little volatile due to the `[Added Associations]` section
that often gets updated when we open a file with a non-default
application (i.e. "Open With" in Thunar's context menu).